@@ -1,21 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <CodeAnalysisRuleSet>../../PlaywrightSharp.ruleset</CodeAnalysisRuleSet>
-  </PropertyGroup>
-  <PropertyGroup>
-    <IsTool>true</IsTool>
-  </PropertyGroup>
-  <Import Project="../../Common/PackageInfo.props" />
-  <Import Project="../../Common/Dependencies.props" />
-  <ItemGroup>
-    <PackageReference Include="CommandLineParser" Version="2.9.0-preview1" />
-    <PackageReference Include="System.CodeDom" Version="5.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\PlaywrightSharp\PlaywrightSharp.csproj" />
-    <ProjectReference Include="..\PlaywrightSharp.BuildTasks\PlaywrightSharp.BuildTasks.csproj" />
-  </ItemGroup>
+	<PropertyGroup>
+		<OutputType>Exe</OutputType>
+		<TargetFramework>net5.0</TargetFramework>
+		<CodeAnalysisRuleSet>../../PlaywrightSharp.ruleset</CodeAnalysisRuleSet>
+	</PropertyGroup>
+	<PropertyGroup>
+		<IsTool>true</IsTool>
+	</PropertyGroup>
+	<Import Project="../../Common/PackageInfo.props" />
+	<Import Project="../../Common/Dependencies.props" />
+	<ItemGroup>
+		<PackageReference Include="CommandLineParser" Version="2.9.0-preview1" />
+		<PackageReference Include="System.CodeDom" Version="5.0.0" />
+	</ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\PlaywrightSharp.BuildTasks\PlaywrightSharp.BuildTasks.csproj" />
+	</ItemGroup>
 </Project>
