@@ -7,22 +7,22 @@ using Xunit.Abstractions;
 namespace PlaywrightSharp.Tests.Page.Network
 {
     ///<playwright-file>network.spec.js</playwright-file>
-    ///<playwright-describe>Page.setExtraHTTPHeaders</playwright-describe>
+    ///<playwright-describe>Page.setextraHTTPHeaders</playwright-describe>
     [Collection(TestConstants.TestFixtureBrowserCollectionName)]
-    public class PageSetExtraHttpHeadersTests : PlaywrightSharpPageBaseTest
+    public class PageSetextraHTTPHeadersTests : PlaywrightSharpPageBaseTest
     {
         /// <inheritdoc/>
-        public PageSetExtraHttpHeadersTests(ITestOutputHelper output) : base(output)
+        public PageSetextraHTTPHeadersTests(ITestOutputHelper output) : base(output)
         {
         }
 
         ///<playwright-file>network.spec.js</playwright-file>
-        ///<playwright-describe>Page.setExtraHTTPHeaders</playwright-describe>
+        ///<playwright-describe>Page.setextraHTTPHeaders</playwright-describe>
         ///<playwright-it>should work</playwright-it>
         [Fact(Timeout = TestConstants.DefaultTestTimeout)]
         public async Task ShouldWork()
         {
-            await Page.SetExtraHttpHeadersAsync(new Dictionary<string, string>
+            await Page.SetextraHTTPHeadersAsync(new Dictionary<string, string>
             {
                 ["Foo"] = "Bar"
             });
@@ -34,13 +34,13 @@ namespace PlaywrightSharp.Tests.Page.Network
         }
 
         ///<playwright-file>network.spec.js</playwright-file>
-        ///<playwright-describe>Page.setExtraHTTPHeaders</playwright-describe>
+        ///<playwright-describe>Page.setextraHTTPHeaders</playwright-describe>
         ///<playwright-it>should work with redirects</playwright-it>
         [Fact(Timeout = TestConstants.DefaultTestTimeout)]
         public async Task ShouldWorkWithRedirects()
         {
             Server.SetRedirect("/foo.html", "/empty.html");
-            await Page.SetExtraHttpHeadersAsync(new Dictionary<string, string>
+            await Page.SetextraHTTPHeadersAsync(new Dictionary<string, string>
             {
                 ["Foo"] = "Bar"
             });
@@ -52,13 +52,13 @@ namespace PlaywrightSharp.Tests.Page.Network
         }
 
         ///<playwright-file>network.spec.js</playwright-file>
-        ///<playwright-describe>Page.setExtraHTTPHeaders</playwright-describe>
+        ///<playwright-describe>Page.setextraHTTPHeaders</playwright-describe>
         ///<playwright-it>should work with extra headers from browser context</playwright-it>
         [Fact(Timeout = TestConstants.DefaultTestTimeout)]
         public async Task ShouldWorkWithExtraHeadersFromBrowserContext()
         {
             await using var context = await Browser.NewContextAsync();
-            await context.SetExtraHttpHeadersAsync(new Dictionary<string, string>
+            await context.SetextraHTTPHeadersAsync(new Dictionary<string, string>
             {
                 ["Foo"] = "Bar"
             });
@@ -71,20 +71,20 @@ namespace PlaywrightSharp.Tests.Page.Network
         }
 
         ///<playwright-file>network.spec.js</playwright-file>
-        ///<playwright-describe>Page.setExtraHTTPHeaders</playwright-describe>
+        ///<playwright-describe>Page.setextraHTTPHeaders</playwright-describe>
         ///<playwright-it>should override extra headers from browser context</playwright-it>
         [Fact(Timeout = TestConstants.DefaultTestTimeout)]
         public async Task ShouldOverrideExtraHeadersFromBrowserContext()
         {
             await using var context = await Browser.NewContextAsync();
-            await context.SetExtraHttpHeadersAsync(new Dictionary<string, string>
+            await context.SetextraHTTPHeadersAsync(new Dictionary<string, string>
             {
                 ["fOo"] = "bAr",
                 ["baR"] = "foO",
             });
             var page = await context.NewPageAsync();
 
-            await page.SetExtraHttpHeadersAsync(new Dictionary<string, string>
+            await page.SetextraHTTPHeadersAsync(new Dictionary<string, string>
             {
                 ["Foo"] = "Bar"
             });
@@ -97,7 +97,7 @@ namespace PlaywrightSharp.Tests.Page.Network
         }
 
         ///<playwright-file>network.spec.js</playwright-file>
-        ///<playwright-describe>Page.setExtraHTTPHeaders</playwright-describe>
+        ///<playwright-describe>Page.setextraHTTPHeaders</playwright-describe>
         ///<playwright-it>should throw for non-string header values</playwright-it>
         [Fact(Skip = "We don't need this test")]
         public void ShouldThrowForNonStringHeaderValues() { }
