@@ -47,7 +47,7 @@ namespace PlaywrightSharp.Tests.Popup
         {
             await using var context = await Browser.NewContextAsync(new BrowserContextOptions
             {
-                ExtraHttpHeaders = new Dictionary<string, string>
+                ExtraHTTPHeaders = new Dictionary<string, string>
                 {
                     ["foo"] = "bar"
                 }
