@@ -75,7 +75,7 @@ namespace PlaywrightSharp
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
@@ -99,7 +99,7 @@ namespace PlaywrightSharp
                 ColorScheme = colorScheme,
                 Locale = locale,
                 IgnoreHTTPSErrors = ignoreHTTPSErrors,
-                ExtraHttpHeaders = extraHttpHeaders,
+                ExtraHTTPHeaders = extraHTTPHeaders,
                 RecordHar = recordHar,
                 RecordVideo = recordVideo,
                 Proxy = proxy,
@@ -124,7 +124,7 @@ namespace PlaywrightSharp
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
@@ -147,7 +147,7 @@ namespace PlaywrightSharp
                 IgnoreHTTPSErrors = ignoreHTTPSErrors,
                 ColorScheme = colorScheme,
                 Locale = locale,
-                ExtraHttpHeaders = extraHttpHeaders,
+                ExtraHTTPHeaders = extraHTTPHeaders,
                 RecordHar = recordHar,
                 RecordVideo = recordVideo,
                 Proxy = proxy,
@@ -182,7 +182,7 @@ namespace PlaywrightSharp
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
@@ -206,7 +206,7 @@ namespace PlaywrightSharp
                 IgnoreHTTPSErrors = ignoreHTTPSErrors,
                 ColorScheme = colorScheme,
                 Locale = locale,
-                ExtraHttpHeaders = extraHttpHeaders,
+                ExtraHTTPHeaders = extraHTTPHeaders,
                 RecordHar = recordHar,
                 RecordVideo = recordVideo,
                 Proxy = proxy,
@@ -231,7 +231,7 @@ namespace PlaywrightSharp
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
@@ -254,7 +254,7 @@ namespace PlaywrightSharp
                 IgnoreHTTPSErrors = ignoreHTTPSErrors,
                 ColorScheme = colorScheme,
                 Locale = locale,
-                ExtraHttpHeaders = extraHttpHeaders,
+                ExtraHTTPHeaders = extraHTTPHeaders,
                 RecordHar = recordHar,
                 RecordVideo = recordVideo,
                 Proxy = proxy,
