@@ -133,7 +133,7 @@ namespace PlaywrightSharp
             bool? acceptDownloads = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             bool? chromiumSandbox = null,
             bool? handleSIGINT = null,
             bool? handleSIGTERM = null,
@@ -173,7 +173,7 @@ namespace PlaywrightSharp
                     AcceptDownloads = acceptDownloads,
                     ColorScheme = colorScheme,
                     Locale = locale,
-                    ExtraHttpHeaders = extraHttpHeaders,
+                    ExtraHTTPHeaders = extraHTTPHeaders,
                     ChromiumSandbox = chromiumSandbox,
                     HandleSIGHUP = handleSIGHUP,
                     HandleSIGINT = handleSIGINT,
@@ -212,7 +212,7 @@ namespace PlaywrightSharp
             bool? acceptDownloads = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             bool? chromiumSandbox = null,
             bool? handleSIGINT = null,
             bool? handleSIGTERM = null,
@@ -251,7 +251,7 @@ namespace PlaywrightSharp
                     AcceptDownloads = acceptDownloads,
                     ColorScheme = colorScheme,
                     Locale = locale,
-                    ExtraHttpHeaders = extraHttpHeaders,
+                    ExtraHTTPHeaders = extraHTTPHeaders,
                     ChromiumSandbox = chromiumSandbox,
                     HandleSIGHUP = handleSIGHUP,
                     HandleSIGINT = handleSIGINT,
