@@ -50,7 +50,7 @@ namespace PlaywrightSharp.Chromium
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
@@ -73,7 +73,7 @@ namespace PlaywrightSharp.Chromium
                 ignoreHTTPSErrors,
                 colorScheme,
                 locale,
-                extraHttpHeaders,
+                extraHTTPHeaders,
                 recordHar,
                 recordVideo,
                 proxy,
@@ -97,7 +97,7 @@ namespace PlaywrightSharp.Chromium
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
@@ -119,7 +119,7 @@ namespace PlaywrightSharp.Chromium
                 ignoreHTTPSErrors,
                 colorScheme,
                 locale,
-                extraHttpHeaders,
+                extraHTTPHeaders,
                 recordHar,
                 recordVideo,
                 proxy,
