@@ -94,7 +94,7 @@ namespace PlaywrightSharp.Chromium
             bool? acceptDownloads = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             bool? chromiumSandbox = null,
             bool? handleSIGINT = null,
             bool? handleSIGTERM = null,
@@ -131,7 +131,7 @@ namespace PlaywrightSharp.Chromium
                 acceptDownloads,
                 colorScheme,
                 locale,
-                extraHttpHeaders,
+                extraHTTPHeaders,
                 chromiumSandbox,
                 handleSIGINT,
                 handleSIGTERM,
@@ -169,7 +169,7 @@ namespace PlaywrightSharp.Chromium
             bool? acceptDownloads = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             bool? chromiumSandbox = null,
             bool? handleSIGINT = null,
             bool? handleSIGTERM = null,
@@ -205,7 +205,7 @@ namespace PlaywrightSharp.Chromium
                 acceptDownloads,
                 colorScheme,
                 locale,
-                extraHttpHeaders,
+                extraHTTPHeaders,
                 chromiumSandbox,
                 handleSIGINT,
                 handleSIGTERM,
