@@ -52,7 +52,7 @@ namespace PlaywrightSharp.Chromium
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
@@ -76,7 +76,7 @@ namespace PlaywrightSharp.Chromium
             bool? ignoreHTTPSErrors = null,
             ColorScheme? colorScheme = null,
             string locale = null,
-            Dictionary<string, string> extraHttpHeaders = null,
+            Dictionary<string, string> extraHTTPHeaders = null,
             RecordHarOptions recordHar = null,
             RecordVideoOptions recordVideo = null,
             ProxySettings proxy = null,
