@@ -1,11 +1,17 @@
-using System.Collections.Generic;
-
 namespace PlaywrightSharp.BuildTasks.Models.Api
 {
     public class PlaywrightMember
     {
-        public string Kind { get; set; }
+        public PlaywrightMemberKind Kind { get; set; }
 
-        public Dictionary<string, PlaywrightArgument> Args { get; set; }
+        public string Name { get; set; }
+
+        public bool Deprecated { get; set; }
+
+        public PlaywrightType Type { get; set; }
+
+        public PlaywrightMember[] Args { get; set; }
+
+        public PlaywrightLangs Langs { get; set; }
     }
 }
